infrastructure/agents/platform/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx[http2]==0.28.1
pydantic==2.10.5
pydantic-settings==2.6.1
python-frontmatter==1.1.0
psycopg[binary,pool]==3.2.3
redis==5.2.1
python-dotenv==1.0.1
sse-starlette==2.1.3
apscheduler==3.10.4
lxml==5.3.0
pyyaml==6.0.2
